# bam: Use libdeflate instead of the standard compress/{gzip,flate}