It results in 2x to 4x performance improvements.

github.com/biogo/hts/bgzf

Before:
  BenchmarkWrite-56              1        2087464279 ns/op
  BenchmarkRead-56               1        4452965285 ns/op

After:
  BenchmarkWrite-56    	       1	1069021742 ns/op
  BenchmarkRead-56     	       1	1195341566 ns/op

Hardware: 56 * Xeon E5-2690@2.60GHz, 256GiB memory.