@yasushi-saito 2018-12-12
It addresses the issue

https://github.com/biogo/hts/issues/124

(Most of the work was done by Chris, by the way)