@kortschak 2018-12-12
I will be able to have a look at this in a few days.