@shenwei356 2018-12-16
Maybe you can use pgzip, I use it a lot.

    gzip "github.com/klauspost/pgzip"


 > Go parallel gzip compression/decompression. This is a fully gzip compatible drop in replacement for "compress/gzip".