@kortschak 2018-12-16
Yes, that's a possibility, though bgzf already uses parallelism. The alternative is "github.com/klauspost/compress/gzip" which has optimisations that I don't believe are in std.