@yasushi-saito 2018-12-17
On Sat, Dec 15, 2018 at 6:54 PM Dan Kortschak <notifications@github.com>
wrote:

> *@kortschak* commented on this pull request.
>
> This can't go in as it is. The condition of adding a Cgo dependency was
> that it was conditional on build tags and off by default
> <https://github.com/biogo/hts/issues/124#issuecomment-401517682>. This is
> always on.
>
libdeflate itself is conditional, so it builds on both cgo and pure
environments (in latter, it just uses compress/flate). Does that not
suffice?


> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/biogo/hts/pull/129#pullrequestreview-185373215>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AAhVsw52SQrIFyfJuiA11jOpGOqx1F7Vks5u5bXNgaJpZM4ZQP7c>
> .
>


-- 
Yasushi Saito