@yasushi-saito 2018-12-17
On Sat, Dec 15, 2018 at 9:32 PM Dan Kortschak <notifications@github.com>
wrote:

> Yes, that's a possibility, though bgzf already uses parallelism. The
> alternative is "github.com/klauspost/compress/gzip" which has
> optimisations that I don't believe are in std.
>
We've used klauspost/compress internally but it provides only minor
improvements over the standard packages - most of its improvements have
been upstreamed already.

> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/biogo/hts/pull/129#issuecomment-447619069>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AAhVs1w5y7zEMCTuTWqPpTL-rOvgSa4Rks5u5drtgaJpZM4ZQP7c>
> .
>


-- 
Yasushi Saito