@kortschak 2018-12-17
OK. So I'll have to look at libdeflate. That will take some time.

From a quick look though, the fact that libdeflate is part of grailbio/base means that all those other packages are pulled just to get libdeflate. I'd be happier if it were in its own repo.

Also, there should be additional tests elements in the traivs matrix to check the Cgo and pure-Go implementation and documentation in bgzf to explain how to get the different builds.

Finally, I see that you have code that is derivative of Klaus' code in libdeflate.go, this should have headers showing Go Authors and Klaus' copyrights on that in addition to Grail's (also the Go and Klaus' LICENSE files).

> libdeflate itself is conditional, so it builds on both cgo and pure environments (in latter, it just uses compress/flate). Does that not suffice?